Code refractored according to new json format
Changed the strategy from multi instance of req performer behaviour to single instance to handle all orders

+++ src/main/java/zarvis/bakery/models/Location.java
@@ -2,15 +2,15 @@ package zarvis.bakery.models;
 
 public class Location {
 	
-	private int x;
-	private int y;
-	public int getX() {
+	private float x;
+	private float y;
+	public float getX() {
 		return x;
 	}
 	public void setX(int x) {
 		this.x = x;
 	}
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 	public void setY(int y) {
