+++ src/main/java/zarvis/bakery/models/Product.java
@@ -30,11 +30,11 @@ public class Product {
 			
 	private int breads_per_box;	
 			
-	private int production_cost;
+	private float production_cost;
 			
-	private int sales_price	;
+	private float sales_price;
 
-	public int getDough_prep_time() {
+	public float getDough_prep_time() {
 		return dough_prep_time;
 	}
 
@@ -106,19 +106,19 @@ public class Product {
 		this.breads_per_box = breads_per_box;
 	}
 
-	public int getProduction_cost() {
+	public float getProduction_cost() {
 		return production_cost;
 	}
 
-	public void setProduction_cost(int production_cost) {
+	public void setProduction_cost(float production_cost) {
 		this.production_cost = production_cost;
 	}
 
-	public int getSales_price() {
+	public float getSales_price() {
 		return sales_price;
 	}
 
-	public void setSales_price(int sales_price) {
+	public void setSales_price(float sales_price) {
 		this.sales_price = sales_price;
 	}
